Add api as package and rename utils
